chore: rename to src
